Fixed dependencies for App.js broken by move to containers folder.
