+++ app/reducer.js
@@ -22,6 +22,13 @@ function queueSonglist(state = [], action) {
   }
 }
 
+function historySonglist(state = [], action) {
+    switch (action.type) {
+      default:
+        return state;
+    }
+}
+
 const partyqApp = combineReducers({
   queueSonglist
 });
